+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated SaxonHE to 12.5
 - Updated sfm-csv to 9.0.2
 - Updated slf4j to 2.0.12
-- Updated testcontainers to 1.19.8
+- Updated testcontainers to 1.20.3
 - Log name of subclass in log messages of Record.
 - `Record.get(<reference>)` now returns a `RecordValue` object which wraps the actual result. 
   It also provides methods to check for an empty (null) value or an error. This is a breaking change to the API.
@@ -30,6 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed dependency on xlsx-streamer
 - Removed direct dependency on tika-parsers-standard-package
 - Removed dependency on javax.activation-api
+
+### Changed
+- `Record.get(<reference>)` now returns a `RecordValue` object which wraps the actual result.
+  It also provides methods to check for an empty (null) value or an error. This is a breaking change to the API.
 
 ## [1.3.1] - 2024-08-12